+++ packages/dds/tree/src/test/feature-libraries/modular-schema/genericFieldKind.spec.ts
@@ -34,7 +34,7 @@ import { singleJsonCursor } from "../../../domains/index.js";
 // eslint-disable-next-line import/no-internal-modules
 import { RebaseRevisionMetadata } from "../../../feature-libraries/modular-schema/index.js";
 import { ValueChangeset, valueField, valueHandler } from "./basicRebasers.js";
-import { testSnapshots } from "./genericField.spec.js";
+import { testSnapshots } from "./genericFieldSnapshots.test.js";
 
 const valueFieldKey: FieldKey = brand("Value");
